+++ src/lib/asyncapiFile.js
@@ -1,9 +1,9 @@
 import { readFile } from 'fs/promises'
 import asyncapi from '@asyncapi/parser'
-import { getConstants } from './constants.js'
+import { getConfigs } from './configs.js'
 
 export async function getParsedAsyncAPI() {
-  const { ASYNCAPI_FILE_PATH } = getConstants()
+  const { ASYNCAPI_FILE_PATH } = getConfigs()
   const asyncapiFileContent = await readFile(ASYNCAPI_FILE_PATH, 'utf-8')
   return asyncapi.parse(asyncapiFileContent)
 }
